Revert "Diffable data source v2"
