+++ MultiSectionCompositionalLayout/Modules/Main/MainExtensions/ViewController + UICollectionViewDataSource.swift
@@ -7,39 +7,41 @@
 
 import UIKit
 
-//MARK: - Diffable data source -
-extension ViewController {
-    func setDataSource() {
-        self.dataSource =
-        UICollectionViewDiffableDataSource<ListSection, ListItem>(
-            collectionView: self.collectionView,
-            cellProvider: {
-                (collectionView, indexPath, itemIdentifier) in
-                let section = self.sections[indexPath.section]
-                return CompositionalLayoutHelper
-                    .getCellFor(section,
-                                collectionView,
-                                indexPath)
-                
-            })
-        self.dataSource.supplementaryViewProvider = {
-            (collectionView, kind, indexPath)
-            -> UICollectionReusableView? in
-            let reusableView = CompositionalLayoutHelper
-                .getReusableViewFor(collectionView,
-                                    of: kind, indexPath,
-                                    self.sections)
-            return reusableView
+extension ViewController: UICollectionViewDelegate, UICollectionViewDataSource {
+    func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return sections.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return sections[section].count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        switch sections[indexPath.section] {
+        case .stories(let items):
+            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "StoryCollectionViewCell", for: indexPath) as! StoryCollectionViewCell
+            cell.setup(items[indexPath.row])
+            return cell
+        case .popular(let items):
+            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PortraitCollectionViewCell", for: indexPath) as! PortraitCollectionViewCell
+            cell.setup(items[indexPath.row])
+            return cell
+        case .comingSoon(let items):
+            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "LandscapeCollectionViewCell", for: indexPath) as! LandscapeCollectionViewCell
+            cell.setup(items[indexPath.row])
+            return cell
         }
     }
     
-    func setSnapshot() {
-        var snapshot = NSDiffableDataSourceSnapshot<ListSection,
-                                                    ListItem>()
-        snapshot.appendSections(self.sections)
-        self.sections.forEach({
-            snapshot.appendItems($0.items, toSection: $0)
-        })
-        self.dataSource.apply(snapshot)
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        switch kind {
+        case UICollectionView.elementKindSectionHeader:
+            let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "CollectionViewHeaderReusableView", for: indexPath) as! CollectionViewHeaderReusableView
+            header.setup(sections[indexPath.section].title)
+            return header
+        default:
+            return UICollectionReusableView()
+        }
     }
 }
+
