+++ MultiSectionCompositionalLayout/Modules/Main/MainHelpers/CompositionalLayoutHelper.swift
@@ -198,44 +198,3 @@ extension CompositionalLayoutHelper {
     
     
 }
-
-
-//MARK: - Diffable data source helper -
-extension CompositionalLayoutHelper {
-    static func getCellFor(_ section: ListSection,
-                           _ collectionView: UICollectionView,
-                           _ indexPath: IndexPath)
-    -> UICollectionViewCell {
-        switch section {
-        case .stories(let items):
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "StoryCollectionViewCell", for: indexPath) as! StoryCollectionViewCell
-            cell.setup(items[indexPath.item])
-            return cell
-        case .popular(let items):
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PortraitCollectionViewCell", for: indexPath) as! PortraitCollectionViewCell
-            cell.setup(items[indexPath.item])
-            return cell
-        case .comingSoon(let items):
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "LandscapeCollectionViewCell", for: indexPath) as! LandscapeCollectionViewCell
-            cell.setup(items[indexPath.item])
-            return cell
-            
-        default: fatalError("something wrong with layer helper")
-        }
-    }
-
-    static func getReusableViewFor(_ collectionView: UICollectionView,
-                            of kind: String,
-                            _ indexPath: IndexPath,
-                            _ sections: [ListSection]) -> UICollectionReusableView? {
-        guard let supView = collectionView
-            .dequeueReusableSupplementaryView(
-                ofKind: kind,
-                withReuseIdentifier: "CollectionViewHeaderReusableView",
-                for: indexPath) as? CollectionViewHeaderReusableView
-        else { return nil }
-        supView.setup(sections[indexPath.section].title)
-        
-        return supView
-    }
-}
